refactor(voice): split realtime.ts along its consumer boundaries
The file grouped six unrelated slices under a vendor name. Move them into
protocol, tools, credentials, context, and voice-settings modules, and
collapse the cloned context envelope and session-update paths while there.
