+++ packages/sidecar-core/src/index.ts
@@ -10,7 +10,11 @@ export * from "./json";
 export * from "./motion-tokens";
 export * from "./provider-mark-paths";
 export * from "./providers";
-export * from "./realtime";
+export * from "./realtime-context";
+export * from "./realtime-credentials";
+export * from "./realtime-protocol";
+export * from "./realtime-tools";
+export * from "./realtime-voice-settings";
 export * from "./session";
 export * from "./session-display";
 export * from "./session-notices";
