+++ packages/sidecar-core/src/realtime-tools.ts
@@ -1,0 +1,875 @@
+import {
+  APP_PANEL_TAB,
+  APP_SETTING_KIND,
+  type AppGuideSetting,
+  type AppGuideSnapshot,
+  type AppPanelTab,
+  appGuideSetting,
+  appToggleValue,
+  FEEDBACK_COMPOSER_KIND,
+  type FeedbackComposerKind,
+  isAppPanelTab,
+  isFeedbackComposerKind,
+  isSessionListSort,
+  SESSION_LIST_SORT,
+  type SessionListSort,
+} from "./guide";
+import {
+  type IssueIdentity,
+  type IssueTransition,
+  issueCommentText,
+  type TrackedIssue,
+} from "./issues";
+import { isRecord } from "./json";
+import {
+  maximumWorkspaceNameLength,
+  type ObservedWorkspaceProject,
+  WORKSPACE_TASK_SUPPORT,
+  type WorkspaceAgentModels,
+  type WorkspaceAgentSelection,
+  workspaceNameText,
+} from "./providers";
+import { maximumFeedbackDraftLength, type RealtimeFunctionCall } from "./realtime-protocol";
+import {
+  type NormalizedSession,
+  SESSION_LOCATION,
+  type SessionControl,
+  type SessionIdentity,
+  sessionMessageText,
+  supportsSessionControl,
+} from "./session";
+
+/**
+ * The acts Luke can carry for the developer, named as Realtime tools. The
+ * session trio are the same acts the panel's rows offer — the two writes, and
+ * the press that opens a session where its provider keeps it — and the issue
+ * pair are the two acts a connected tracker takes. All run the same gauntlet:
+ * a call is validated against the observed roster before anything leaves the
+ * renderer, and the main process validates it again against what it observed
+ * before anything happens. Luke is another way to ask, never a wider one.
+ *
+ * Creating a workspace is the one act with no row yet to mirror, and it keeps
+ * the same posture: a call can only name a project its provider reported on
+ * the latest observation pass — sent to the conversation as [workspace
+ * projects] the way the roster is — and the main process validates it again
+ * against what its adapters actually offered.
+ *
+ * The last three are the same presses turned toward the app itself: a settings
+ * change goes through the bridge call the setting's own row uses, validated
+ * against the app guide first; showing the panel is the capsule's press with a
+ * tab chosen out loud; and opening the feedback composer is the tray item's
+ * press, carrying at most the developer's own words as a starting draft — it
+ * can never send, because sending is the composer's own button, pressed by
+ * hand. None of the three reaches a provider.
+ */
+export const REALTIME_TOOL = {
+  SEND_SESSION_MESSAGE: "send_session_message",
+  RUN_SESSION_CONTROL: "run_session_control",
+  OPEN_SESSION: "open_session",
+  CREATE_WORKSPACE: "create_workspace",
+  ADD_WORKSPACE_AGENT: "add_workspace_agent",
+  UPDATE_ISSUE_STATE: "update_issue_state",
+  COMMENT_ON_ISSUE: "comment_on_issue",
+  CHANGE_APP_SETTING: "change_app_setting",
+  SHOW_PANEL: "show_panel",
+  OPEN_FEEDBACK_COMPOSER: "open_feedback_composer",
+} as const;
+
+export type RealtimeToolName = (typeof REALTIME_TOOL)[keyof typeof REALTIME_TOOL];
+
+const SESSION_TOOL_NAMES: ReadonlySet<string> = new Set([
+  REALTIME_TOOL.SEND_SESSION_MESSAGE,
+  REALTIME_TOOL.RUN_SESSION_CONTROL,
+  REALTIME_TOOL.OPEN_SESSION,
+  // A workspace ask names a provider's project rather than a session, but it
+  // is the same family of act — carried by the session carrier, validated in
+  // sessionToolAction against the projects the conversation was shown.
+  REALTIME_TOOL.CREATE_WORKSPACE,
+  REALTIME_TOOL.ADD_WORKSPACE_AGENT,
+]);
+
+const ISSUE_TOOL_NAMES: ReadonlySet<string> = new Set([
+  REALTIME_TOOL.UPDATE_ISSUE_STATE,
+  REALTIME_TOOL.COMMENT_ON_ISSUE,
+]);
+
+/** Whether a tool call names one of the two session acts. */
+export function isSessionToolName(name: string): boolean {
+  return SESSION_TOOL_NAMES.has(name);
+}
+
+/** Whether a tool call names one of the two issue acts. */
+export function isIssueToolName(name: string): boolean {
+  return ISSUE_TOOL_NAMES.has(name);
+}
+
+const SESSION_IDENTITY_PARAMETERS = {
+  provider_id: {
+    type: "string",
+    description: "The provider_id of the session, exactly as the roster lists it.",
+  },
+  provider_session_id: {
+    type: "string",
+    description: "The provider_session_id of the session, exactly as the roster lists it.",
+  },
+} as const;
+
+const ISSUE_IDENTITY_PARAMETERS = {
+  tracker_id: {
+    type: "string",
+    description: "The tracker_id of the issue, exactly as the issue roster lists it.",
+  },
+  issue_id: {
+    type: "string",
+    description:
+      "The issue_id of the issue, exactly as the issue roster lists it, such as LUKE-123.",
+  },
+} as const;
+
+/** The tool schemas a Realtime session is configured with. */
+export function realtimeToolDefinitions(): readonly Record<string, unknown>[] {
+  return [
+    {
+      type: "function",
+      name: REALTIME_TOOL.SEND_SESSION_MESSAGE,
+      description:
+        "Send a message the developer just asked you to send to one observed session. " +
+        "Only sessions the roster marks as taking messages can receive one.",
+      parameters: {
+        type: "object",
+        properties: {
+          ...SESSION_IDENTITY_PARAMETERS,
+          text: {
+            type: "string",
+            description: "The message, in the developer's own words or their clear intent.",
+          },
+        },
+        required: ["provider_id", "provider_session_id", "text"],
+      },
+    },
+    {
+      type: "function",
+      name: REALTIME_TOOL.RUN_SESSION_CONTROL,
+      description:
+        "Run a control one observed session advertises, such as stopping its current run. " +
+        "Only controls the roster lists for that session exist.",
+      parameters: {
+        type: "object",
+        properties: {
+          ...SESSION_IDENTITY_PARAMETERS,
+          control_id: {
+            type: "string",
+            description: "The control's id, exactly as the roster lists it in parentheses.",
+          },
+        },
+        required: ["provider_id", "provider_session_id", "control_id"],
+      },
+    },
+    {
+      type: "function",
+      name: REALTIME_TOOL.OPEN_SESSION,
+      description:
+        "Open one observed session on the developer's screen, the same as pressing its row. " +
+        "Only sessions the roster marks as able to be opened have somewhere to open.",
+      parameters: {
+        type: "object",
+        properties: { ...SESSION_IDENTITY_PARAMETERS },
+        required: ["provider_id", "provider_session_id"],
+      },
+    },
+    {
+      type: "function",
+      name: REALTIME_TOOL.CREATE_WORKSPACE,
+      description:
+        "Create a new workspace — a new agent — the developer just asked for, in one project " +
+        "a provider listed. Only projects the [workspace projects] context lists exist.",
+      parameters: {
+        type: "object",
+        properties: {
+          provider_id: {
+            type: "string",
+            description: "The provider_id of the project, exactly as the projects list gives it.",
+          },
+          project_id: {
+            type: "string",
+            description: "The project_id, exactly as the projects list gives it.",
+          },
+          name: {
+            type: "string",
+            description:
+              "A short name for the workspace, only when the developer chose one; " +
+              "the provider names it otherwise.",
+          },
+          task: {
+            type: "string",
+            description:
+              "What the developer asked the new agent to work on, in their own words or their " +
+              "clear intent. Required where the projects list says a task is needed; omitted " +
+              "where it says the project takes none.",
+          },
+          model: {
+            type: "string",
+            description:
+              "The model for the new agent, exactly as the app guide's model setting lists it, " +
+              "only when the developer named one for this creation; the settings decide otherwise.",
+          },
+          effort: {
+            type: "string",
+            description:
+              "The effort level riding that model, exactly as the guide lists it, only when the " +
+              "developer named both; never without a model.",
+          },
+        },
+        required: ["provider_id", "project_id"],
+      },
+    },
+    {
+      type: "function",
+      name: REALTIME_TOOL.ADD_WORKSPACE_AGENT,
+      description:
+        "Start another agent in the workspace one observed session runs in. Only sessions " +
+        "whose roster entry lists new agents can take one, only as an agent kind it lists.",
+      parameters: {
+        type: "object",
+        properties: {
+          ...SESSION_IDENTITY_PARAMETERS,
+          agent: {
+            type: "string",
+            description: "The kind of agent, exactly as the roster lists it under new agents.",
+          },
+          name: {
+            type: "string",
+            description:
+              "A short name for the new agent's session, only when the developer chose one.",
+          },
+          task: {
+            type: "string",
+            description:
+              "What the developer asked the new agent to work on, in their own words or their " +
+              "clear intent, when they gave it something to start on.",
+          },
+          model: {
+            type: "string",
+            description:
+              "The model for the new agent, exactly as the app guide's model setting lists it, " +
+              "only when the developer named one for this agent; the settings decide otherwise.",
+          },
+          effort: {
+            type: "string",
+            description:
+              "The effort level riding that model, exactly as the guide lists it, only when the " +
+              "developer named both; never without a model.",
+          },
+        },
+        required: ["provider_id", "provider_session_id", "agent"],
+      },
+    },
+    {
+      type: "function",
+      name: REALTIME_TOOL.UPDATE_ISSUE_STATE,
+      description:
+        "Move one tracked issue to a state the developer just asked for. " +
+        "Only issues the issue roster lists exist, and only the states it lists for one.",
+      parameters: {
+        type: "object",
+        properties: {
+          ...ISSUE_IDENTITY_PARAMETERS,
+          state: {
+            type: "string",
+            description: "The target state's name, exactly as the issue roster lists it.",
+          },
+        },
+        required: ["tracker_id", "issue_id", "state"],
+      },
+    },
+    {
+      type: "function",
+      name: REALTIME_TOOL.COMMENT_ON_ISSUE,
+      description:
+        "Add a comment the developer just asked you to add to one tracked issue. " +
+        "Only issues the issue roster marks as taking comments can receive one.",
+      parameters: {
+        type: "object",
+        properties: {
+          ...ISSUE_IDENTITY_PARAMETERS,
+          body: {
+            type: "string",
+            description: "The comment, in the developer's own words or their clear intent.",
+          },
+        },
+        required: ["tracker_id", "issue_id", "body"],
+      },
+    },
+    {
+      type: "function",
+      name: REALTIME_TOOL.CHANGE_APP_SETTING,
+      description:
+        "Change one of Luke's own settings the developer just asked for. " +
+        "Only settings the app guide marks as changeable by voice can be changed.",
+      parameters: {
+        type: "object",
+        properties: {
+          setting_id: {
+            type: "string",
+            description: "The setting_id, exactly as the app guide lists it.",
+          },
+          value: {
+            type: "string",
+            description:
+              "The new value: on or off for a switch, or one of the choices the guide lists.",
+          },
+        },
+        required: ["setting_id", "value"],
+      },
+    },
+    {
+      type: "function",
+      name: REALTIME_TOOL.SHOW_PANEL,
+      description:
+        "Show Luke's own panel on the developer's screen, the same as pressing the capsule — or, " +
+        "when the panel is already open, switch it to the named tab, the same as pressing that tab. " +
+        "It can open the sessions list — narrowed to one provider or location, ordered by urgency or recency — or the settings tab.",
+      parameters: {
+        type: "object",
+        properties: {
+          tab: {
+            type: "string",
+            enum: Object.values(APP_PANEL_TAB),
+            description: "Which tab to open or switch to. Defaults to sessions.",
+          },
+          filter: {
+            type: "string",
+            description:
+              "Narrows the session list: all, local, cloud, or the provider_id of one observed provider. Only meaningful on the sessions tab.",
+          },
+          sort: {
+            type: "string",
+            enum: Object.values(SESSION_LIST_SORT),
+            description:
+              "Reorders the session list: urgency puts what needs the developer first, recency puts what moved last first. Only meaningful on the sessions tab.",
+          },
+        },
+        required: [],
+      },
+    },
+    {
+      type: "function",
+      name: REALTIME_TOOL.OPEN_FEEDBACK_COMPOSER,
+      description:
+        "Open the composer for a note the developer sends the founders by hand. " +
+        "It opens and may draft; it never sends — the developer edits and presses Send themselves.",
+      parameters: {
+        type: "object",
+        properties: {
+          kind: {
+            type: "string",
+            enum: Object.values(FEEDBACK_COMPOSER_KIND),
+            description:
+              "What the note is: feedback about Luke, or a prompt for the founders. A refused ask offered onward is a prompt.",
+          },
+          draft: {
+            type: "string",
+            description:
+              "Optional starting text: the developer's own ask, in their words. Never words they did not say.",
+          },
+        },
+        required: ["kind"],
+      },
+    },
+  ];
+}
+
+/** What one validated tool call asks for, ready for the bridge that carries it. */
+export type SessionToolAction =
+  | { kind: "message"; identity: SessionIdentity; text: string }
+  | { kind: "control"; identity: SessionIdentity; control: SessionControl }
+  | { kind: "open"; identity: SessionIdentity }
+  | {
+      kind: "create-workspace";
+      providerId: string;
+      providerProjectId: string;
+      name?: string;
+      task?: string;
+      /** The model the developer named for this one creation, resolved to ids. */
+      agentSelection?: WorkspaceAgentSelection;
+    }
+  | {
+      kind: "add-agent";
+      identity: SessionIdentity;
+      agent: string;
+      name?: string;
+      task?: string;
+      /** The model the developer named for this one agent, as its wire id. */
+      model?: string;
+      /** The effort riding that model, when the developer named both. */
+      effort?: string;
+    }
+  | { kind: "refused"; reason: string };
+
+function textArgument(record: Record<string, unknown>, key: string): string | undefined {
+  const value = record[key];
+  const normalized = typeof value === "string" ? value.trim() : "";
+  return normalized || undefined;
+}
+
+/**
+ * Resolves a model the developer named — by the label the guide lists it
+ * under, or its id — to the wire pairing an endpoint takes, held to the
+ * build's documented entries for the provider. The effort, when named, must
+ * be one the resolved model's own agent documents: the pairing is validated
+ * as the whole it will be sent as.
+ */
+function resolveWorkspaceAgentModel(
+  entries: readonly WorkspaceAgentModels[],
+  modelWord: string,
+  effortWord: string | undefined,
+): { selection: WorkspaceAgentSelection } | { refused: string } {
+  const normalizedModel = modelWord.trim().toLowerCase();
+  const named = entries
+    .flatMap((entry) => entry.models.map((model) => ({ entry, model })))
+    .find(
+      ({ model }) =>
+        model.label.toLowerCase() === normalizedModel || model.id.toLowerCase() === normalizedModel,
+    );
+  if (!named) return { refused: "No documented model goes by that name here." };
+  let effort: string | undefined;
+  if (effortWord !== undefined) {
+    const normalizedEffort = effortWord.trim().toLowerCase();
+    effort = named.entry.efforts.find((candidate) => candidate.toLowerCase() === normalizedEffort);
+    if (!effort) {
+      return {
+        refused:
+          named.entry.efforts.length > 0
+            ? `That model's effort is one of ${named.entry.efforts.join(", ")}.`
+            : "That model takes no effort level.",
+      };
+    }
+  }
+  return {
+    selection: { agent: named.entry.agent, model: named.model.id, ...(effort ? { effort } : {}) },
+  };
+}
+
+/**
+ * Validates one tool call against the sessions actually being observed. This
+ * is the renderer's half of the gauntlet — the main process re-validates
+ * against its registry — and it exists so a call the model composed can only
+ * name a session Luke was shown, doing something that session advertised.
+ * Everything else is refused with a reason Luke can say aloud.
+ */
+export function sessionToolAction(
+  call: RealtimeFunctionCall,
+  sessions: readonly NormalizedSession[],
+  workspaceProjects: readonly ObservedWorkspaceProject[] = [],
+  // The models a creation ask may name, per provider — the app's own
+  // build-documented tables, handed in so this stays brand-neutral. The
+  // default offers none, so an ask that names a model is refused rather than
+  // forwarded unchecked.
+  agentModels: (providerId: string) => readonly WorkspaceAgentModels[] = () => [],
+): SessionToolAction {
+  let parsed: unknown;
+  try {
+    parsed = JSON.parse(call.argumentsJson);
+  } catch {
+    return { kind: "refused", reason: "The tool call's arguments were not readable." };
+  }
+  if (!isRecord(parsed)) {
+    return { kind: "refused", reason: "The tool call's arguments were not readable." };
+  }
+
+  // A creation ask names a project rather than a session, so it is validated
+  // against the projects the conversation was shown before the roster is even
+  // consulted — the same discipline, against the list that actually offered it.
+  if (call.name === REALTIME_TOOL.CREATE_WORKSPACE) {
+    const project = workspaceProjects.find(
+      (candidate) =>
+        candidate.providerId === textArgument(parsed, "provider_id") &&
+        candidate.providerProjectId === textArgument(parsed, "project_id"),
+    );
+    if (!project) {
+      return { kind: "refused", reason: "No listed project matches that identity." };
+    }
+    let name: string | undefined;
+    if (parsed.name !== undefined) {
+      name = workspaceNameText(parsed.name);
+      if (!name) {
+        return {
+          kind: "refused",
+          reason: `A workspace name has to be under ${maximumWorkspaceNameLength} characters and longer than nothing.`,
+        };
+      }
+    }
+    // The task is held to the project's own word for it: a project that takes
+    // none cannot be handed one, a project that needs one cannot be created
+    // without it, and the text itself is bounded like the message it is.
+    let task: string | undefined;
+    if (parsed.task !== undefined) {
+      if (project.taskSupport === WORKSPACE_TASK_SUPPORT.NONE) {
+        return { kind: "refused", reason: "That project takes no opening task." };
+      }
+      task = sessionMessageText(parsed.task);
+      if (!task) {
+        return {
+          kind: "refused",
+          reason: "A task has to be shorter than a document and longer than nothing.",
+        };
+      }
+    } else if (project.taskSupport === WORKSPACE_TASK_SUPPORT.REQUIRED) {
+      return {
+        kind: "refused",
+        reason: "That project needs an opening task to create a workspace.",
+      };
+    }
+    // A model named for this one creation resolves against the provider's own
+    // documented table, and the effort only ever rides a model: alone it has
+    // nothing documented to attach to.
+    const spokenModel = textArgument(parsed, "model");
+    const spokenEffort = textArgument(parsed, "effort");
+    if (spokenEffort !== undefined && spokenModel === undefined) {
+      return { kind: "refused", reason: "An effort rides a model; name the model too." };
+    }
+    let agentSelection: WorkspaceAgentSelection | undefined;
+    if (spokenModel !== undefined) {
+      const resolved = resolveWorkspaceAgentModel(
+        agentModels(project.providerId),
+        spokenModel,
+        spokenEffort,
+      );
+      if ("refused" in resolved) return { kind: "refused", reason: resolved.refused };
+      agentSelection = resolved.selection;
+    }
+    return {
+      kind: "create-workspace",
+      providerId: project.providerId,
+      providerProjectId: project.providerProjectId,
+      ...(name ? { name } : {}),
+      ...(task ? { task } : {}),
+      ...(agentSelection ? { agentSelection } : {}),
+    };
+  }
+
+  const providerId = textArgument(parsed, "provider_id");
+  const providerSessionId = textArgument(parsed, "provider_session_id");
+  const session = sessions.find(
+    (candidate) =>
+      candidate.providerId === providerId && candidate.providerSessionId === providerSessionId,
+  );
+  if (!session) {
+    return { kind: "refused", reason: "No observed session matches that identity." };
+  }
+  const identity: SessionIdentity = {
+    providerId: session.providerId,
+    providerSessionId: session.providerSessionId,
+  };
+
+  if (call.name === REALTIME_TOOL.SEND_SESSION_MESSAGE) {
+    if (!session.canReceiveMessage) {
+      return { kind: "refused", reason: "That session does not take messages right now." };
+    }
+    const text = sessionMessageText(parsed.text);
+    if (!text) {
+      return {
+        kind: "refused",
+        reason: "A message has to be shorter than a document and longer than nothing.",
+      };
+    }
+    return { kind: "message", identity, text };
+  }
+
+  if (call.name === REALTIME_TOOL.RUN_SESSION_CONTROL) {
+    const controlId = textArgument(parsed, "control_id");
+    const control = session.controls.find((candidate) => candidate.id === controlId);
+    if (!controlId || !control || !supportsSessionControl(session, controlId)) {
+      return { kind: "refused", reason: "That session advertises no such control." };
+    }
+    return { kind: "control", identity, control };
+  }
+
+  if (call.name === REALTIME_TOOL.OPEN_SESSION) {
+    // The action carries the identity, never the address: the main process
+    // reads the link back out of its own registry, the same as a pressed row.
+    if (!session.detail.link) {
+      return { kind: "refused", reason: "That session has no address to open." };
+    }
+    return { kind: "open", identity };
+  }
+
+  if (call.name === REALTIME_TOOL.ADD_WORKSPACE_AGENT) {
+    // The agent must be one this session's own roster entry listed: the list
+    // is the provider's word for what its endpoint takes, so an ask outside it
+    // is refused rather than forwarded to be refused.
+    const agent = textArgument(parsed, "agent");
+    if (!agent || !session.spawnableAgents.includes(agent)) {
+      return { kind: "refused", reason: "That session lists no such agent to add." };
+    }
+    let name: string | undefined;
+    if (parsed.name !== undefined) {
+      name = workspaceNameText(parsed.name);
+      if (!name) {
+        return {
+          kind: "refused",
+          reason: `A session name has to be under ${maximumWorkspaceNameLength} characters and longer than nothing.`,
+        };
+      }
+    }
+    let task: string | undefined;
+    if (parsed.task !== undefined) {
+      task = sessionMessageText(parsed.task);
+      if (!task) {
+        return {
+          kind: "refused",
+          reason: "A task has to be shorter than a document and longer than nothing.",
+        };
+      }
+    }
+    // A model named for this one agent resolves within the asked-for kind
+    // alone: the developer's chosen agent is never re-decided by the model
+    // they named beside it, so a mismatch is a refusal rather than a swap.
+    const spokenModel = textArgument(parsed, "model");
+    const spokenEffort = textArgument(parsed, "effort");
+    if (spokenEffort !== undefined && spokenModel === undefined) {
+      return { kind: "refused", reason: "An effort rides a model; name the model too." };
+    }
+    let selection: WorkspaceAgentSelection | undefined;
+    if (spokenModel !== undefined) {
+      const entries = agentModels(session.providerId).filter(
+        (candidate) => candidate.agent === agent,
+      );
+      const resolved = resolveWorkspaceAgentModel(entries, spokenModel, spokenEffort);
+      if ("refused" in resolved) {
+        return {
+          kind: "refused",
+          reason: resolved.refused.startsWith("No documented model")
+            ? `A ${agent} agent runs no model by that name.`
+            : resolved.refused,
+        };
+      }
+      selection = resolved.selection;
+    }
+    return {
+      kind: "add-agent",
+      identity,
+      agent,
+      ...(name ? { name } : {}),
+      ...(task ? { task } : {}),
+      ...(selection ? { model: selection.model } : {}),
+      ...(selection?.effort ? { effort: selection.effort } : {}),
+    };
+  }
+
+  return { kind: "refused", reason: "No such tool exists." };
+}
+
+/** What one validated issue tool call asks for, ready for the bridge that carries it. */
+export type IssueToolAction =
+  | { kind: "issue-state"; identity: IssueIdentity; transition: IssueTransition }
+  | { kind: "issue-comment"; identity: IssueIdentity; body: string }
+  | { kind: "refused"; reason: string };
+
+/**
+ * Validates one issue tool call against the issues actually observed. The
+ * renderer's half of the same gauntlet the session tools run — the main
+ * process re-validates against what it observed — so a call the model
+ * composed can only name an issue Luke was shown, going somewhere its
+ * tracker advertised. Everything else is refused with a reason Luke can say
+ * aloud.
+ */
+export function issueToolAction(
+  call: RealtimeFunctionCall,
+  issues: readonly TrackedIssue[],
+): IssueToolAction {
+  let parsed: unknown;
+  try {
+    parsed = JSON.parse(call.argumentsJson);
+  } catch {
+    return { kind: "refused", reason: "The tool call's arguments were not readable." };
+  }
+  if (!isRecord(parsed)) {
+    return { kind: "refused", reason: "The tool call's arguments were not readable." };
+  }
+
+  const trackerId = textArgument(parsed, "tracker_id");
+  const issueId = textArgument(parsed, "issue_id");
+  const issue = issues.find(
+    (candidate) => candidate.trackerId === trackerId && candidate.identifier === issueId,
+  );
+  if (!issue) {
+    return { kind: "refused", reason: "No tracked issue matches that identity." };
+  }
+  const identity: IssueIdentity = {
+    trackerId: issue.trackerId,
+    identifier: issue.identifier,
+  };
+
+  if (call.name === REALTIME_TOOL.UPDATE_ISSUE_STATE) {
+    const state = textArgument(parsed, "state");
+    // Spoken names arrive with their case retold rather than copied, so the
+    // match forgives case alone — never spelling — and only while it stays
+    // unambiguous. Two advertised states apart only in case are not a guess
+    // Luke gets to make.
+    const named = state
+      ? issue.transitions.filter(
+          (candidate) => candidate.name.toLowerCase() === state.toLowerCase(),
+        )
+      : [];
+    const transition =
+      named.find((candidate) => candidate.name === state) ??
+      (named.length === 1 ? named[0] : undefined);
+    if (!transition) {
+      return { kind: "refused", reason: "That issue lists no such state." };
+    }
+    return { kind: "issue-state", identity, transition };
+  }
+
+  if (call.name === REALTIME_TOOL.COMMENT_ON_ISSUE) {
+    if (!issue.canComment) {
+      return { kind: "refused", reason: "That issue does not take comments." };
+    }
+    const body = issueCommentText(parsed.body);
+    if (!body) {
+      return {
+        kind: "refused",
+        reason: "A comment has to be shorter than a document and longer than nothing.",
+      };
+    }
+    return { kind: "issue-comment", identity, body };
+  }
+
+  return { kind: "refused", reason: "No such tool exists." };
+}
+
+/**
+ * The whole-list scope a spoken panel ask may name. The rest of the filter
+ * vocabulary is not this module's to define: a location is a session's own
+ * `location`, and a provider is its `provider_id`, so a spoken filter is
+ * validated against the observed roster rather than against a second list.
+ */
+export const SESSION_LIST_ALL = "all";
+
+/**
+ * What one validated app tool call asks for, ready for the app to perform.
+ * The feedback action opens the composer and nothing else: `draft` is at most
+ * the developer's own words, placed only into an empty note, and what the
+ * composer holds leaves only by its own Send button — no action here sends.
+ */
+export type AppToolAction =
+  | { kind: "setting"; setting: AppGuideSetting; value: string }
+  | { kind: "panel"; tab: AppPanelTab; filter?: string; sort?: SessionListSort }
+  | { kind: "feedback"; composer: FeedbackComposerKind; draft?: string }
+  | { kind: "refused"; reason: string };
+
+/** Whether a tool call is about the app itself rather than about a session. */
+export function isAppToolCall(call: RealtimeFunctionCall): boolean {
+  return (
+    call.name === REALTIME_TOOL.CHANGE_APP_SETTING ||
+    call.name === REALTIME_TOOL.SHOW_PANEL ||
+    call.name === REALTIME_TOOL.OPEN_FEEDBACK_COMPOSER
+  );
+}
+
+/**
+ * Validates the value a spoken change carries against the setting it names.
+ * A toggle takes the guide's own two words (and their unambiguous synonyms);
+ * a choice takes exactly one of the values the guide listed. Anything else is
+ * refused with the accepted set, so the refusal is also the correction.
+ */
+function appSettingValue(setting: AppGuideSetting, value: unknown): string | undefined {
+  if (setting.kind === APP_SETTING_KIND.TOGGLE) return appToggleValue(value);
+  const normalized = typeof value === "string" ? value.trim().toLowerCase() : "";
+  return setting.choices?.find((choice) => choice.toLowerCase() === normalized);
+}
+
+/**
+ * Validates a spoken session-list filter against the sessions actually being
+ * observed. A filter that would show nothing is refused rather than applied:
+ * the panel would quietly fall back to showing everything, and Luke would have
+ * reported a narrowing that never happened.
+ */
+function panelFilterAction(
+  filter: string,
+  sessions: readonly NormalizedSession[],
+): { filter: string } | { reason: string } {
+  if (filter === SESSION_LIST_ALL) return { filter };
+  if (filter === SESSION_LOCATION.LOCAL || filter === SESSION_LOCATION.CLOUD) {
+    if (sessions.some((session) => session.location === filter)) return { filter };
+    return { reason: `No ${filter} sessions are observed right now.` };
+  }
+  if (sessions.some((session) => session.providerId === filter)) return { filter };
+  return { reason: "No observed session belongs to that provider." };
+}
+
+/**
+ * Validates one app tool call against the guide the app actually provided and
+ * the sessions actually observed. The same posture as {@link sessionToolAction}:
+ * a call the model composed can only name a setting the guide lists, changing
+ * it to a value the guide accepts, a panel view the roster can fill, or the
+ * composer on one of its own two kinds — and a setting the guide marks as
+ * by-hand-only is refused with the path to it, so the refusal Luke voices is
+ * itself the guidance.
+ */
+export function appToolAction(
+  call: RealtimeFunctionCall,
+  guide: AppGuideSnapshot,
+  sessions: readonly NormalizedSession[],
+): AppToolAction {
+  let parsed: unknown;
+  try {
+    parsed = JSON.parse(call.argumentsJson);
+  } catch {
+    return { kind: "refused", reason: "The tool call's arguments were not readable." };
+  }
+  if (!isRecord(parsed)) {
+    return { kind: "refused", reason: "The tool call's arguments were not readable." };
+  }
+
+  if (call.name === REALTIME_TOOL.CHANGE_APP_SETTING) {
+    const setting = appGuideSetting(guide, textArgument(parsed, "setting_id"));
+    if (!setting) {
+      return { kind: "refused", reason: "The app guide lists no such setting." };
+    }
+    if (!setting.adjustable) {
+      return {
+        kind: "refused",
+        reason: `${setting.label} can only be changed by hand: ${setting.manual}`,
+      };
+    }
+    const value = appSettingValue(setting, parsed.value);
+    if (value === undefined) {
+      const accepted =
+        setting.kind === APP_SETTING_KIND.TOGGLE ? "on or off" : (setting.choices ?? []).join(", ");
+      return { kind: "refused", reason: `${setting.label} takes ${accepted}.` };
+    }
+    return { kind: "setting", setting, value };
+  }
+
+  if (call.name === REALTIME_TOOL.OPEN_FEEDBACK_COMPOSER) {
+    const composer = parsed.kind;
+    if (!isFeedbackComposerKind(composer)) {
+      return { kind: "refused", reason: "The composer writes feedback or a prompt, nothing else." };
+    }
+    // The draft is the developer's ask restated, so it is bounded like a typed
+    // one; a blank draft is no draft, and the composer simply opens empty.
+    const draft = textArgument(parsed, "draft")?.slice(0, maximumFeedbackDraftLength);
+    return { kind: "feedback", composer, ...(draft ? { draft } : {}) };
+  }
+
+  if (call.name === REALTIME_TOOL.SHOW_PANEL) {
+    const tab = parsed.tab ?? APP_PANEL_TAB.SESSIONS;
+    if (!isAppPanelTab(tab)) {
+      return { kind: "refused", reason: "The panel has no such tab." };
+    }
+    const sort = textArgument(parsed, "sort");
+    if (sort !== undefined && !isSessionListSort(sort)) {
+      return { kind: "refused", reason: "The list orders by urgency or by recency." };
+    }
+    const filter = textArgument(parsed, "filter");
+    if (filter === undefined) {
+      return { kind: "panel", tab, ...(sort !== undefined ? { sort } : {}) };
+    }
+    const outcome = panelFilterAction(filter, sessions);
+    if ("reason" in outcome) return { kind: "refused", reason: outcome.reason };
+    return { kind: "panel", tab, filter: outcome.filter, ...(sort !== undefined ? { sort } : {}) };
+  }
+
+  return { kind: "refused", reason: "No such tool exists." };
+}
